pyproject.toml
[project]
name = "mediawiki-to-pdf"
description = "TBD"
license = {text = "MIT" }
version = "0.0"
dependencies = [
    "pdfkit",
	"pyyaml",
    "mediawiki_session @ git+https://github.com/NMRhub/mediawiki_session.git"
]
requires-python= ">= 3.10"
readme = "README.md"
authors = [
    {name = "Gerard"},
    {email = "gweatherby@uchc.edu"}
]

[tool.setuptools]
package-dir = {"" = "src" }  # Tells setuptools that the root package is in the "src" directory

[tool.setuptools.packages.find]
where = ["src"]  # Specify the "src" directory to look for packages


[project.scripts]
mediawiki_to_pdf = "mediawiki_to_pdf.main:main"

[project.optional-dependencies] 
# test = ['pytest']

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[[tool.uv.index]]
name = "nmrbox"
url = "https://apt.nmrbox.org/pip"

